util: Move crypto metrics to their own separate file
